BinarizeEvidence now accept objects that implements the CreateFactor interface. This will allow to define the type of the new node in the graph that will collect and binarize the evidence. Somoe other minor fixes.
